+++ resources/js/package/vite-plugin/index.js
@@ -1,29 +1,13 @@
 import vue from '@vitejs/plugin-vue';
 import { spawn } from 'child_process';
 import tailwindExclusions from './tailwind-exclusions.js';
+import preventServer from './prevent-server.js';
 
 const statamic = function (options) {
     return {
         name: 'statamic',
 
         config(config, { command }) {
-            if (command === 'serve' && !process.env.STATAMIC_FORCE_SERVE) {
-                console.log('\x1b[33m[Statamic] Vite dev server current not supported. Automatically running "vite build --watch" instead...\x1b[0m');
-                console.log('\x1b[90m[Statamic] Use STATAMIC_FORCE_SERVE=1 to bypass this behavior.\x1b[0m');
-
-                const child = spawn('npx', ['vite', 'build', '--watch'], {
-                    stdio: 'inherit',
-                    cwd: process.cwd()
-                });
-
-                child.on('error', (err) => {
-                    console.error('Failed to start vite build --watch:', err);
-                    process.exit(1);
-                });
-
-                process.exit(0);
-            }
-
             // Ensure rollupOptions exists
             config.build = config.build || {};
             config.build.rollupOptions = config.build.rollupOptions || {};
@@ -59,6 +43,7 @@ const statamic = function (options) {
 export default function (options = {}) {
     return [
         statamic(options),
+        preventServer(),
         tailwindExclusions(),
         vue(options.vue || {}),
     ];
